requirements.txt
anyio==4.9.0
certifi==2025.1.31
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
openpyxl==3.1.5
packaging==24.2
peewee==3.17.9
pluggy==1.5.0
pytest==8.3.5
python-telegram-bot==22.0
sniffio==1.3.1
typing_extensions==4.13.1
